+++ packages/vue/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Controlled `modelValue` updates no longer emit `update:modelValue` or `change`.
+
 ## [0.1.0] - 2026-07-14
 
 ### Added